Pass search term into search func
- no longer assume it is the simpleRepresentation, but that is what is still used all the time at this moment.
- laying the ground work for allowing the search term to be editable.
