+++ Sources/MissingArtwork/Model.swift
@@ -38,7 +38,8 @@ public class Model: ObservableObject {
         self.missingArtworks = try await Array(Set<MissingArtwork>(missingArtworks))
 
         for missingArtwork in self.missingArtworks {
-          await fetchImageURLs(missingArtwork: missingArtwork, token: token)
+          await fetchImageURLs(
+            missingArtwork: missingArtwork, term: missingArtwork.simpleRepresentation, token: token)
         }
       } catch {
         debugPrint("Unable to fetch missing artworks: \(error)")
@@ -47,11 +48,27 @@ public class Model: ObservableObject {
     }
   }
 
-  func fetchImageURLs(missingArtwork: MissingArtwork, token: String) async {
+  func searchURL(term: String, limit: Int = 2) -> URL {
+    var urlComponents = URLComponents()
+    urlComponents.scheme = "https"
+    urlComponents.host = "api.music.apple.com"
+    urlComponents.path = "/v1/catalog/us/search"
+    urlComponents.queryItems = [
+      URLQueryItem(name: "term", value: term),
+      URLQueryItem(name: "types", value: "albums"),
+      URLQueryItem(name: "limit", value: "\(limit)"),
+    ]
+    if let url = urlComponents.url {
+      return url
+    }
+    return URL(string: "missing")!  // Use an bogus URL and allow the networking layer return an error.
+  }
+
+  func fetchImageURLs(missingArtwork: MissingArtwork, term: String, token: String) async {
     if self.missingArtworkURLs[missingArtwork] == nil {
       let fetcher = ArtworkURLFetcher(token: token)
       do {
-        self.missingArtworkURLs[missingArtwork] = try await fetcher.fetch(missingArtwork.searchURL)
+        self.missingArtworkURLs[missingArtwork] = try await fetcher.fetch(searchURL(term: term))
       } catch {
         debugPrint("Unable to fetch missing artwork URLs: (\(missingArtwork)) - \(error)")
         self.missingArtworkURLs[missingArtwork] = []
